pyproject.toml
[tool.poetry]
name = "is_len"
version = "0.1.0"
description = "Detect faces in statues, determine if Lenin."
authors = ["Xavier Holt <holt.xavier@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
torch = "^1.10.2"
pytorch-lightning = "^1.5.10"
hydra-core = "^1.1.1"
timm = "^0.5.4"
loguru = "^0.6.0"
pandas = "^1.4.1"
setuptools = "59.5.0"
gdown = "^4.3.1"
transformers = "^4.16.2"
scipy = "^1.8.0"
opencv-python = "^4.5.5"
wandb = "^0.12.10"
albumentations = {git = "https://github.com/albumentations-team/albumentations"}
fire = "^0.4.0"
tqdm = "^4.62.3"
pycocotools = "^2.0.4"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
yapf = "^0.32.0"
ipython = "^8.0.1"
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
